Move selectActiveCycleForDomainLead to its own dependency-free file
test-api CI failed because the picker test imported the domain-lead
route module, which transitively pulls in ~/lib/db → the generated
prisma client. CI's test-api job doesn't run prisma generate before
npm test, so that import fails at module-resolution time.

Other tests work around this by vi.mock'ing ~/lib/db at the top
(hoisted by vitest before any import runs). But the cleaner answer
for a pure helper is to put it somewhere with no db dependency at
all. New file app/hiring/lib/cycle-picker.ts has no imports; the
route and the test both import from there.

No behavior change.
